paging added

+++ TeamProject (Book Reservation)/BL/DTOs/QueryResultDTO.cs
@@ -4,7 +4,7 @@ namespace BL.DTOs
 {
     public class QueryResultDTO<TEntityDTO> where TEntityDTO : class, IEntityDTO
     {
-        public long TotalItemsCount { get; set; }
+        public int TotalItemsCount { get; set; }
         public int? RequestedPageNumber { get; set; }
         public int PageSize { get; set; }
         public IEnumerable<TEntityDTO> Items { get; set; }
