+++ TeamProject (Book Reservation)/Infrastructure/Query/QueryResult.cs
@@ -5,7 +5,7 @@ namespace Infrastructure.Query
 {
     public class QueryResult<TEntity> where TEntity : class, IEntity
     {
-        public long TotalItemsCount { get; set; }
+        public int TotalItemsCount { get; set; }
         public int? RequestedPageNumber { get; set; }
         public int PageSize { get; set; }
         public IEnumerable<TEntity> Items { get; set; }
